rename: service 관련 파일들 service 패키지로 이동
서비스 패키지의 위치로 인하여 홈페이지에서 빌드가 안됨
