+++ src/main/java/bridge/Application.java
@@ -1,5 +1,7 @@
 package bridge;
 
+import bridge.service.BridgeGame;
+import bridge.service.PlayerSetting;
 import bridge.view.InputView;
 import bridge.view.OutputView;
