+++ src/main/java/bridge/service/BridgeGame.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.service;
 
 import bridge.domain.GameData;
 import bridge.dto.HistoryMapDto;
