+++ src/main/java/bridge/service/BridgeRandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.service;
 
 import camp.nextstep.edu.missionutils.Randoms;
