+++ src/main/java/bridge/service/PlayerCommend.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.service;
 
 public enum PlayerCommend {
     MOVE_UPPER("U"),
