+++ src/main/java/bridge/service/PlayerSetting.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.service;
 
 import bridge.domain.Player;
 import bridge.dto.BridgeCorrectLocationDto;
